+++ src/useRecord.ts
@@ -6,5 +6,13 @@ export function useRecord<T, K extends string, C extends Config<K, T>>(
   store: CrudStore<T, K, C, ValidatedConfig<K, T, C>>
 ): { [key: string]: T } | null {
   const data = store((s) => s.data);
-  return useMemo(() => data ? Object.fromEntries(data) : null, [data]);
+  const dk = store.config.detailKey;
+  return useMemo(() => {
+    if (!data) return null;
+    const rec: { [key: string]: T } = {};
+    data.forEach((item) => {
+      rec[String((item as any)[dk])] = item;
+    });
+    return rec;
+  }, [data]);
 }